Move API Connectors CRUD inline into its Setup Hub card, add deep link
Replaces the standalone /connectors page with the CRUD UI embedded
directly in the ConnectorsSection card's expandable body, matching the
existing convention (CameraSection/BridgeSection/StorageSection all
work this way — a standalone page only exists for the two that
predate the Setup Hub redesign). Adds /setup/connectors as a deep
link: it renders SetupHubPage with the card pre-expanded and scrolled
into view via useRoute, so other pages can still link straight to it
without duplicating the UI on its own route.
