+++ packages/lcyt-web/src/main.jsx
@@ -58,7 +58,6 @@ const ProductionVisualPage   = lazyImport(() => import('./components/ProductionV
 const PlannerPage            = lazyImport(() => import('./components/PlannerPage').then(m => ({ default: m.PlannerPage })));
 const TranslationsPage       = lazyImport(() => import('./components/TranslationsPage').then(m => ({ default: m.TranslationsPage })));
 const AiSettingsPage         = lazyImport(() => import('./components/AiSettingsPage').then(m => ({ default: m.AiSettingsPage })));
-const ConnectorsPage         = lazyImport(() => import('./components/ConnectorsPage').then(m => ({ default: m.ConnectorsPage })));
 const AdminUsersPage         = lazyImport(() => import('./components/AdminUsersPage').then(m => ({ default: m.AdminUsersPage })));
 const AdminUserDetailPage    = lazyImport(() => import('./components/AdminUserDetailPage').then(m => ({ default: m.AdminUserDetailPage })));
 const AdminProjectsPage      = lazyImport(() => import('./components/AdminProjectsPage').then(m => ({ default: m.AdminProjectsPage })));
@@ -202,7 +201,6 @@ function SidebarApp() {
             <Route path="/planner" component={PlannerPage} />
             <Route path="/translations" component={TranslationsPage} />
             <Route path="/ai" component={AiSettingsPage} />
-            <Route path="/connectors" component={ConnectorsPage} />
             <Route path="/admin/users/:id" component={AdminUserDetailPage} />
             <Route path="/admin/users" component={AdminUsersPage} />
             <Route path="/admin/projects/:key" component={AdminProjectDetailPage} />
@@ -216,6 +214,7 @@ function SidebarApp() {
             <Route path="/team" component={TeamPage} />
             <Route path="/setup" component={SetupHubPage} />
             <Route path="/setup/wizard" component={SetupWizardPage} />
+            <Route path="/setup/connectors" component={SetupHubPage} />
             <Route path="/account" component={AccountPage} />
             <Route path="/settings" component={SettingsPage} />
             {/* Legacy URL aliases */}
