+++ packages/plugins/lcyt-connectors/CLAUDE.md
@@ -33,10 +33,10 @@ app.use('/variables', createVariablesRouter(db, auth, bus, engine));
 - `src/lib/metacode-variables.js` — `interpolateVariables(text, snapshot)`: client-side `{{name}}` insertion, mirrors `interpolate.js` above.
 - `src/hooks/useVariables.js` — `useVariables({ backendUrl, connected, getToken })`: fetches the initial snapshot (`GET /variables`), subscribes to `GET /variables/events` SSE, exposes `snapshot()` and `refresh(connectorSlug, requestSlug, waitMs?)`.
 - `src/components/InputBar.jsx` — pointer-change effect fires pointer-tier triggers once and (re)starts/stops the prefetch-tier's background interval (default 3000ms — the frontend doesn't fetch the request's configured `prefetch_interval_ms` before starting the loop, so it uses the plan's stated default); `doSend()` fires send-tier triggers fire-and-forget, does one `waitMs`-bearing prefetch-tier top-up call, resolves `{{ }}` in the outgoing text via `interpolateVariables`, and merges `variables.snapshot()` into `codes` **additively** alongside the existing localStorage-backed `getActiveCodes()` (manual codes still win — this preserves `ActionsPanel`/`QuickActionsPopover` behavior unchanged rather than migrating them onto the new backend, which remains follow-on work per the plan's own scoping).
+- `src/components/setup-hub/ConnectorsSection.jsx` — the Connector/Request/response-mapping/Variable management UI itself: a Setup Hub card with full CRUD inline in its expandable body, same convention as `CameraSection`/`BridgeSection`/`StorageSection` (no standalone page — see root `CLAUDE.md`'s Metacode Organization / Setup Hub conventions). Deep-linkable via `/setup/connectors`, which renders `SetupHubPage` with this card pre-expanded and scrolled into view (`SetupHubPage`'s `useRoute('/setup/connectors')`).
 
 ## Known gaps / follow-on work (see plan §9, §10)
 
-- Standalone Connector/Request/Variable management UI (new pages) — not built; CRUD is API-only today.
 - `ActionsPanel`/`QuickActionsPopover` still write to `localStorage` (`metacode-active.js`), not to `POST/PUT /variables` — migrating them is separate frontend work the plan itself scopes independently.
 - No "stale/default used" (`variableFallbacks`) visible indicator on sent captions yet (plan §4's open UI question).
 - `{{ }}` insertion is only wired into `InputBar.jsx`'s outgoing caption text, not yet into planner/DSK template text fields (plan §10.3).
